chore: move page layout to folder
