+++ src/pages/floating-video.tsx
@@ -1,7 +1,7 @@
 import { graphql, PageProps } from 'gatsby';
 
 import { Video } from '../components/content/floating-video/Video';
-import PageLayout from '../components/layout/PageLayout';
+import PageLayout from '../components/layout/page-layout/PageLayout';
 import SEO from '../components/layout/SEO';
 import MockVideo from '../images/video/ocean_960.mp4';
