+++ src/pages/image.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 
 import { keyframes, styled } from '../../stitches.config';
-import PageLayout from '../components/layout/PageLayout';
+import PageLayout from '../components/layout/page-layout/PageLayout';
 
 export default function ImageTipsPage() {
   const [loaded, setLoaded] = useState(false);
